start separating plots
